requirements.txt
requests
bs4
python-dateutil
pandas
lxml
tqdm